Move consent removal to Consent service
Previously this was added to the deprovision service. That fits well but
is not semantically correct. As this is not a deprovision
(user-lifecycle) action but a user triggerd profile api call.
