+++ src/OpenConext/EngineBlock/Service/DeprovisionServiceInterface.php
@@ -32,6 +32,4 @@ interface DeprovisionServiceInterface
      * @param CollabPersonId $id
      */
     public function delete(CollabPersonId $id);
-
-    public function deleteOneConsentFor(CollabPersonId $id, string $serviceProviderEntityId): bool;
 }
